test: extract render functions to separate module
