+++ vitest.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+  test: {
+    exclude: [
+      "**/node_modules/**",
+      "**/.reference/**",
+      "**/dist/**",
+    ],
+  },
+})
